+++ src/main/java/br/com/dbccompany/sistemadereembolso/Sistema/de/reembolso/entity/UsuarioEntity.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.Set;
 
@@ -30,7 +31,7 @@ public class UsuarioEntity implements UserDetails {
     private String senha;
 
     @Column(name = "valor_total")
-    private Double valorTotal;
+    private BigDecimal valorTotal;
 
     @Column(name = "status")
     private Boolean status;
